Fix pause semantics.
Turns out, we couldn't reliably interrupt KVM in the kernel
simply by setting stepping on or off. It's somewhat difficult
to get around the Go scheduler in this case, so we need to isolate
some critical pieces inside C extensions to avoid thread switching.

This leads to some fairly complex-seeming pause-related code, but
in the end I believe that this is necessary.

Change-Id: I1709df4ec7fea6827f38f6f984c48e5ab8dbe241