Refactor state and fix serialization issues.
Many of the device objects had issues with serialization.

This commit addresses most of those issues (implementing a custom
marshal or unmarshal method where necessary) and refactors the internal
state. Instead of using a separate devices and vcpus object, we now use
just a single state object which contains both devices and vcpus.

Change-Id: I9ee1f1852f85365863e1dae22aa7269c0f1967f3