Implement zero-downtime upgrades.
As a mechanism for providing zero-downtime upgrades, as well as exercising
built-in suspend/resume mechanisms we support hitting novmm with a SIGHUP
and causing a re-exec() of the process with the current state.

Note that this required that we change the way state is passed in to the
novmm process. In order to ensure that we can use arbitrarily-sized state,
we now accept an FD instead of a bloating the command line.

Change-Id: Ie526e7bb96eb3037a3e9e190d5b4d299e3ffb68e