Control refactor: split types into Rpc and Control.
It was getting a bit frustrating having to ensure that
all public methods of the Control matched the rpc spec.

This commit splits the control types into the server and
an RPC object, which the actual control calls hang from.
This makes the separate very clear (and ensures that we
won't actually expose methods that shouldn't be exposed).

Change-Id: I10a826a3ae21df9d8d3115555f27633e07b5b0e3