Implement a special restart mode.
This allows us to effectively keep VMs running and upgrade KVM.
While this probably isn't super useful without a larger management
framework, this at least gives us a compelling use case for the
SIGHUP restart functionality (which was largely introduced for
simple testing and verification of all the state serialization).

This also moves signal catching as close as possible to the start
of the program, in order to reduce the chance of missing them.
(This sort of sucks, but I don't really want to go plubming into
the depths of the runtime in order to fix this problem).

Change-Id: Id1c7d778107cd51d2aa6d65c8db81b3bd4de7b5a