Working eventfd implementation.
Since this works, I've also removed the flag that allows you
to enable or disable eventfds. This prevents a bounce out of the
kernel to handle device emulation (i.e. notifying the virtio
backend that a new buffer has been added), so should improve
latency decently. There's no solid reason to disable it.

Change-Id: I4656b30c2ee18fe3e287dd99d4afc38ae4f1e9da