Initial wstat() implementation.
This at least allows touch (and rename, almost) to work.

All of these operations have become quite complex in commit,rollback
logic. At some point in the future, it would make sense to create an
interface for all the Operations to implement, along the lines of:

   Prepare(), Commit() | Rollback()

Then the parse methods would simply return an interface object which
would be called appropriately. Anyways, this is work for a future
commit.

Change-Id: I153e9378f2e16ca02bef1d7e183f2c8efba7e41d